VIH-9395 - moved constants to DAL helpers - refactor return object for empty list - removed unused code
