+++ BookingsApi/BookingsApi.DAL/Services/HearingService.cs
@@ -10,8 +10,6 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using BookingsApi.Common.Const;
-using BookingsApi.DAL.Queries;
 
 namespace BookingsApi.DAL.Services
 {
