+++ BookingsApi/BookingsApi/Startup.cs
@@ -17,7 +17,6 @@ using Microsoft.ApplicationInsights.Extensibility;
 using Microsoft.Extensions.Hosting;
 using BookingsApi.Contract.Configuration;
 using BookingsApi.Common.Services;
-using BookingsApi.Services;
 
 namespace BookingsApi
 {
